src/api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.1
greenlet==3.1.0
h11==0.14.0
idna==3.8
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic-core==2.23.3
pyjwt==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
sqlalchemy==2.0.34
starlette==0.38.5
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
